requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@985f5732c0ad720e5008ec36019bec6fc15cfc3a
distlib==0.3.4
filelock==3.6.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pip==20.3.4
platformdirs==2.5.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
requests==2.27.1
setuptools==60.10.0
six==1.16.0
urllib3==1.26.9
virtualenv==20.13.4
wheel==0.37.1
yarl==1.7.2
youtube-dl==2021.12.17
pillow==9.0.1
humanize
black
jishaku @ git+https://github.com/Gorialis/jishaku
